+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.6.10
+ - add support for Json in mssql
+ - fix decoding MySQL DECIMAL values as f64 (used to provoke a panic)
+ - add support for MSSQL decimal and numeric types
+
 ## 0.6.9
  - mssql: improved chrono date and time support, fix bug with timezeone decoding
  - MySQL: add support for DateTime<FixedOffset>